[Prover] Rename state.rs to prover_state.rs
