+++ prover-service/src/lib.rs
@@ -13,8 +13,8 @@ pub mod load_vk;
 pub mod metrics;
 pub mod prover_config;
 pub mod prover_key;
+pub mod prover_state;
 pub mod proving;
-pub mod state;
 pub mod training_wheels;
 pub mod utils;
 pub mod witness_gen;
