+++ prover-service/src/main.rs
@@ -11,7 +11,7 @@ use clap::Parser;
 use http::Method;
 use prover_service::prover_config::ProverServiceConfig;
 use prover_service::prover_key::TrainingWheelsKeyPair;
-use prover_service::{state::*, *};
+use prover_service::{prover_state::*, *};
 use std::{net::SocketAddr, sync::Arc, time::Duration};
 use tower::ServiceBuilder;
 use tower_http::cors::{Any, CorsLayer};
