+++ prover-service/src/tests/common/mod.rs
@@ -11,7 +11,7 @@ use crate::{
     handlers::prove_handler,
     jwk_fetching::{KeyID, DECODING_KEY_CACHE},
     prover_config::ProverServiceConfig,
-    state::ProverServiceState,
+    prover_state::ProverServiceState,
 };
 use aptos_crypto::{
     ed25519::{Ed25519PrivateKey, Ed25519PublicKey},
